添加了评论和回复的异常处理，修复了一些bug,修改了表结构

+++ src/main/java/com/chen/communit/controller/PublishController.java
@@ -33,7 +33,7 @@ public class PublishController {
             @RequestParam(value = "title",required = false) String title,
             @RequestParam(value = "description",required = false) String description,
             @RequestParam(value = "tag",required = false) String tag,
-            @RequestParam(value="id",required = false) Integer id,
+            @RequestParam(value="id",required = false) Long id,
             HttpServletRequest request,
             Model model
     ) {
@@ -78,7 +78,7 @@ public class PublishController {
 
     //修改问题时显示出原来的数据
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable("id") Integer id,
+    public String edit(@PathVariable("id") Long id,
                        Model model){
         QuestionDTO question = questionService.getById(id);
         model.addAttribute("title",question.getTitle());
