+++ src/main/java/com/chen/communit/controller/QuestionController.java
@@ -21,7 +21,7 @@ public class QuestionController {
      * @return
      */
     @GetMapping("/question/{id}")
-    public String question(@PathVariable(name = "id") Integer id,
+    public String question(@PathVariable(name = "id") long id,
                            Model model){
         QuestionDTO questionDTO = questionService.getById(id);
         //累加阅读数
@@ -36,7 +36,7 @@ public class QuestionController {
      * @return
      */
     @GetMapping("/delete/{id}")
-    public String delQuestion(@PathVariable("id") Integer id){
+    public String delQuestion(@PathVariable("id") long id){
         questionService.delById(id);
         return "redirect:/profile/questions";
     }
